A HelloWorld.sol change.
It is no longer acceptable to leave out the SPDX-License-Identifier.  Doing so will cause deployment problems; specifically; "HardhatError: HH700: Artifact for contract 'HelloWorld' not found."  I used "// SPDX-License-Identifier: UNLICENSED" and this allowed me to deploy without error.

This change means the note following the compile command has to be changed to remove the line, "You might get a warning about SPDX license identifier not provided in source file, but no need to worry about that...", since it seems you do. :)  I simply changed it to suggest that any problems could be referred to the discord channel.